+++ src/Server/Startup/preflight.ts
@@ -1,6 +1,12 @@
 import jetpaxHome from "../../Common/jetpaxHome"
 import fs from "fs"
 import {execSync} from "child_process"
+import runShutdownTasks from "../Shutdown/runShutdownTasks"
+
+// Register shutdown tasks
+process.on('SIGTERM', async () => {
+    await runShutdownTasks()
+})
 
 // Create the home folder if it hasn't already been created
 if (!fs.existsSync(jetpaxHome)) {